Fix compile error

+++ Kepler-Server/src/main/java/org/alexdev/kepler/dao/mysql/PetDao.java
@@ -98,7 +98,7 @@ public class PetDao {
             row = preparedStatement.executeQuery();
 
             while (row.next()) {
-                petDetails = new PetDetails(row.getInt("id"), row.getLong("item_id"), row.getString("name"),
+                petDetails = new PetDetails(row.getInt("id"), row.getInt("item_id"), row.getString("name"),
                         row.getString("type"), row.getString("race"), row.getString("colour"), row.getInt("nature_positive"),
                         row.getInt("nature_negative"), row.getFloat("friendship"), row.getLong("born"), row.getLong("last_kip"),
                         row.getLong("last_eat"), row.getLong("last_drink"), row.getLong("last_playtoy"), row.getLong("last_playuser"),
